work any type (#1)
* work any type

* change read_data to work with T

* Fixed styles.

+++ src/util/function.hpp
@@ -128,7 +128,7 @@ void Function::deriv_softmax(Tensor<dim1, dim2, dim3, dim4, dim5, T> *t) {
       for (int j = 0; j < row; ++j) {
         int idx = l*(row*col) + k*row + j;
         v[idx] = exp(v[idx]) / sum;
-        v[idx] = v[idx] * (1.0 - Converter::ToDouble(v[idx]));
+        v[idx] = v[idx] * (1.0 - Converter::ToFloat(v[idx]));
       }
     }
   }
